Refactor Docker image build process and add entrypoint script
This commit refactors the Docker image build process in the packer-build.yml file. It introduces changes to the source "docker" "image" section, including adding user actions, setting the working directory to /actions-runner, and specifying the entrypoint as /opt/entrypoint.sh. Additionally, it adds a provisioner file to copy the entrypoint.sh script to the /opt directory.

The buildscript.sh.tpl file is also modified to include additional package installations, create the actions-runner/_work directory, download and extract the actions-runner binary, create the /opt/hostedtoolcache directory, and set permissions for the entrypoint.sh script.

A new entrypoint.sh file is added, which sets the PATH environment variable, configures the GitHub runner, and runs the runner.